moved atom code into separate package
